requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.70
langchain-ollama==0.3.5
langchain-openai==0.3.28
ollama==0.5.1
openai==1.97.0
python-dotenv==1.1.1
pyttsx3==2.99
pytz==2025.2
SpeechRecognition==3.14.3
PyAudio==0.2.14
PyQt6==6.8.1
PyQt6-tools==6.8.0.1
elevenlabs
